New panels with templated AutoRefine3dManager
