+++ src/core/gui_core_headers.h
@@ -114,7 +114,8 @@
 #include "../gui/OverviewPanel.h"
 #include "../gui/AbInitio3DPanel.h"
 #include "../gui/AssetPickerComboPanel.h"
-#include "../gui/AutoRefine3dPanel.h"
+#include "../gui/AutoRefine3dPanelSpa.h"
+#include "../gui/AutoRefine3dPanelRx.h"
 #include "../gui/DisplayRefinementResultsPanel.h"
 #include "../gui/PopupTextDialog.h"
 #include "../gui/LargeAngularPlotDialog.h"
