+++ src/gui/ActionsPanel.cpp
@@ -14,53 +14,10 @@ void ActionsPanel::OnActionsBookPageChanged(wxBookCtrlEvent& event) {
     extern MyFindParticlesPanel* findparticles_panel;
     extern MyRefine2DPanel*      classification_panel;
     extern AbInitio3DPanel*      ab_initio_3d_panel;
-    extern AutoRefine3DPanel*    auto_refine_3d_panel;
+    extern AutoRefine3DPanelSpa* auto_refine_3d_panel_spa;
+    extern AutoRefine3DPanelRx*  auto_refine_3d_panel_rx;
     extern MyRefine3DPanel*      refine_3d_panel;
     extern RefineCTFPanel*       refine_ctf_panel;
     extern Generate3DPanel*      generate_3d_panel;
     extern Sharpen3DPanel*       sharpen_3d_panel;
-
-#ifdef __WXOSX__
-    // Necessary for MacOS to refresh the panels
-    if ( event.GetSelection( ) == 0 ) {
-        align_movies_panel->Layout( );
-        align_movies_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 1 ) {
-        fitctf_panel->Layout( );
-        fitctf_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 2 ) {
-        findparticles_panel->Layout( );
-        findparticles_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 3 ) {
-        classification_panel->Layout( );
-        classification_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 4 ) {
-        ab_initio_3d_panel->Layout( );
-        ab_initio_3d_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 5 ) {
-        auto_refine_3d_panel->Layout( );
-        auto_refine_3d_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 6 ) {
-        refine_3d_panel->Layout( );
-        refine_3d_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 7 ) {
-        refine_ctf_panel->Layout( );
-        refine_ctf_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 8 ) {
-        generate_3d_panel->Layout( );
-        generate_3d_panel->Refresh( );
-    }
-    else if ( event.GetSelection( ) == 9 ) {
-        sharpen_3d_panel->Layout( );
-        sharpen_3d_panel->Refresh( );
-    }
-#endif
 }
