+++ src/gui/AutoRefine3dManager.h
@@ -1,5 +1,5 @@
-#ifndef __AutoRefine3DPanel__
-#define __AutoRefine3DPanel__
+#ifndef __src_gui_AutoRefine3DManager_h__
+#define __src_gui_AutoRefine3DManager_h__
 
 /**
 @file
@@ -8,11 +8,13 @@ Subclass of Refine3DPanel, which is generated by wxFormBuilder.
 
 #include "ProjectX_gui.h"
 
-class AutoRefine3DPanel;
+class AutoRefine3DPanelSpa;
+class AutoRefine3DPanelRx;
 
+template <class MyParentType>
 class AutoRefinementManager {
   public:
-    AutoRefine3DPanel* my_parent;
+    MyParentType* my_parent;
 
     bool     active_should_mask;
     bool     active_should_auto_mask;
@@ -85,9 +87,10 @@ class AutoRefinementManager {
     wxArrayString current_reference_filenames;
     wxArrayLong   current_reference_asset_ids;
 
-    void SetParent(AutoRefine3DPanel* wanted_parent);
+    void SetParent(MyParentType* wanted_parent);
 
     AutoRefinementManager( );
+
     void BeginRefinementCycle( );
     void CycleRefinement( );
 
@@ -115,85 +118,10 @@ class AutoRefinementManager {
 
     void DoMasking( );
 
-    const char* Type( ) const { return "AutoRefinementManager"; };
+    const char* Type( ) const { return "AutoRefine3dManager"; };
 
     //	void StartRefinement();
     //	void StartReconstruction();
-};
-
-class AutoRefine3DPanel : public AutoRefine3DPanelParent {
-    friend class AutoRefinementManager;
-
-  protected:
-    // Handlers for Refine3DPanel events.
-
-    void OnUpdateUI(wxUpdateUIEvent& event);
-    void OnExpertOptionsToggle(wxCommandEvent& event);
-    void OnInfoURL(wxTextUrlEvent& event);
-    void TerminateButtonClick(wxCommandEvent& event);
-    void FinishButtonClick(wxCommandEvent& event);
-    void StartRefinementClick(wxCommandEvent& event);
-    void ResetAllDefaultsClick(wxCommandEvent& event);
-
-    void OnUseMaskCheckBox(wxCommandEvent& event);
-    void OnAutoMaskButton(wxCommandEvent& event);
-
-    // overridden socket methods..
-
-    void OnSocketJobResultMsg(JobResult& received_result);
-    void OnSocketJobResultQueueMsg(ArrayofJobResults& received_queue);
-    void SetNumberConnectedText(wxString wanted_text);
-    void SetTimeRemainingText(wxString wanted_text);
-    void OnSocketAllJobsFinished( );
-
-    int length_of_process_number;
-
-    AutoRefinementManager my_refinement_manager;
-
-    int active_orth_thread_id;
-    int active_mask_thread_id;
-    int next_thread_id;
-
-  public:
-    wxStopWatch stopwatch;
-
-    long time_of_last_result_update;
-
-    bool refinement_package_combo_is_dirty;
-    bool run_profiles_are_dirty;
-    bool volumes_are_dirty;
-
-    JobResult* buffered_results;
-    long       selected_refinement_package;
-
-    //int length_of_process_number;
-
-    JobTracker my_job_tracker;
-
-    bool auto_mask_value; // this is needed to keep track of the automask, as the radiobutton will be overidden to no when masking is selected
-
-    bool running_job;
-
-    AutoRefine3DPanel(wxWindow* parent);
-
-    void Reset( );
-    void SetDefaults( );
-    void SetInfo( );
-
-    void WriteInfoText(wxString text_to_write);
-    void WriteErrorText(wxString text_to_write);
-    void WriteBlueText(wxString text_to_write);
-
-    void FillRefinementPackagesComboBox( );
-    void FillRunProfileComboBoxes( );
-
-    void NewRefinementPackageSelected( );
-
-    void OnRefinementPackageComboBox(wxCommandEvent& event);
-    void OnInputParametersComboBox(wxCommandEvent& event);
-
-    void OnMaskerThreadComplete(wxThreadEvent& my_event);
-    void OnOrthThreadComplete(ReturnProcessedImageEvent& my_event);
 };
 
 #endif
